requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
halo==0.0.31
idna==3.4
log-symbols==0.0.14
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.6
Pygments==2.15.1
PyYAML==6.0
regex==2023.5.5
requests==2.29.0
rich==13.3.5
six==1.16.0
spinners==0.0.24
termcolor==2.3.0
tqdm==4.65.0
urllib3==1.26.15
yarl==1.9.2
